+++ assets/script/books.js
@@ -1,0 +1,5 @@
+const searchInput = document.getElementById("search")
+
+document.getElementById("submit_search").onclick = () => {
+  location.href = `/${doc}/search/${searchInput.value}/1`
+}