+++ assets/script/incAccess.js
@@ -1,10 +1,11 @@
+const
+  paths = location.pathname.split("/"),
+  doc = paths[1] ?? ""
+  book = paths[2] ?? ""
+  chapter = paths[3] ?? ""
+  verse = paths[4] ?? "0"
+
 (async () => {
-  const
-    paths = location.pathname.split("/"),
-    doc = paths[1] ?? ""
-    book = paths[2] ?? ""
-    chapter = paths[3] ?? ""
-    verse = paths[4] ?? "0"
   await fetch("/api/incAccess", {
     method: "post",
     body: JSON.stringify({